--- a/radix-engine/Cargo.toml
+++ b/radix-engine/Cargo.toml
@@ -17,7 +17,7 @@ lru = { version = "0.8.1", default-features = false, optional = true}
 moka = { version = "0.9.4", features = ["sync"], default-features = false, optional = true }
 strum = { version = "0.24", default-features = false, features = ["derive"] }
 perfcnt = { version = "0.8.0", optional = true }
-tikv-jemallocator = { version = "0.5" }
+mimalloc = { version = "0.1.34", default-features = false }
 
 # WASM de-/serialization
 parity-wasm = { version = "0.42.2" }