… assets/blueprints/Cargo.lock
… simulator/Cargo.lock