## Summary

This PR replaces jemalloc with mimalloc for better cross-platform support (the Rust binding).
- [`tikv/jemallocator`](https://github.com/tikv/jemallocator) only supports limited platforms, as stated on the home page;
- [`purpleprotocol/mimalloc_rust`](https://github.com/purpleprotocol/mimalloc_rust) is tested by CI on Ubuntu, MacOS and Windows.
---

Interestingly, contradictory to Microsoft's claim, the performance is slightly worse than jemalloc.

**First CI Run**

Test | Base | PR | %
-- | -- | -- | --
Radiswap::run | 4.9±0.09ms | 5.1±0.09ms | +4.08%
SpinLoop::run | 154.3±5.20ms | 156.9±5.82ms | +1.69%
Transfer::run | 1381.5±21.88µs | 1530.0±55.96µs | +10.75%

**Second CI Run**

Test | Base | PR | %
-- | -- | -- | --
Radiswap::run | 4.9±0.05ms | 5.0±0.02ms | +2.04%
SpinLoop::run | 155.2±2.09ms | 156.1±3.30ms | +0.58%
Transfer::run | 1377.0±13.72µs | 1479.7±25.71µs | +7.46%


**Third CI Run**

Test | Base | PR | %
-- | -- | -- | --
Radiswap::run | 4.9±0.05ms | 5.1±0.05ms | +4.08%
SpinLoop::run | 158.4±2.44ms | 158.3±3.36ms | -0.06%
Transfer::run | 1362.9±9.52µs | 1522.3±15.07µs | +11.70%