@mstrug-rdx 2023-03-13
On default allocator `Transfer::run` bench takes ~1.4751 ms (Linux AWS ec2, 3 runs), so it performs better than `mimalloc`.